requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
charset-normalizer==2.1.1
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-dotenv==1.4.2
django-filter==22.1
django-request-logging==0.7.5
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-dotenv==0.5.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
wincertstore==0.2
